Refactor the utility module
The utilitarian module was divided
into several semantic ones.
