+++ src/utils/common.js
@@ -13,18 +13,6 @@ export const getRandomInteger = (a = 0, b = 1) => {
 };
 
 
-// *** Boolean flag: is property active... ***
-export const isActive = () => {
-  return Boolean(getRandomInteger(0, 1));
-};
-
-
-// *** Function for humanization date ***
-export const humanizeDate = (dateValue, dateFormat) => {
-  return dayjs(dateValue).format(dateFormat);
-};
-
-
 // *** Function for getting a list of random values as a string. ***
 export const getRandomValuesList = (valuesArray, minValueCount, maxValueCount) => {
   // --- Glossary for some special array properties ---
@@ -57,37 +45,7 @@ export const getRandomValuesList = (valuesArray, minValueCount, maxValueCount) =
 };
 
 
-// *** Glossary for element positions ***
-export const RenderPosition = {
-  BEFOREEND: 'beforeend',
-  AFTEREND: 'afterend',
-};
-
-
-// *** Function for rendering elements ***
-export const renderElement = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-
-    case RenderPosition.AFTEREND:
-      container.after(element);
-      break;
-  }
-};
-
-
-// *** Function for rendering components to site page ***
-export const renderTemplate = (container, template, elementPlace) => {
-  container.insertAdjacentHTML(elementPlace, template);
-};
-
-
-// *** Function for generating DOM-elements ***
-export const generateElement = (template) => {
-  const newElement = document.createElement('div');
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
+// *** Function for humanization date ***
+export const humanizeDate = (dateValue, dateFormat) => {
+  return dayjs(dateValue).format(dateFormat);
 };
